# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CashCraft API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfileAvatarStage =
  (typeof UserProfileAvatarStage)[keyof typeof UserProfileAvatarStage];

export const UserProfileAvatarStage = {
  novice: "novice",
  apprentice: "apprentice",
  journeyman: "journeyman",
  master: "master",
  legend: "legend",
} as const;

export interface UserProfile {
  id: string;
  name: string;
  level: number;
  xp: number;
  xpToNextLevel: number;
  futureScore: number;
  streak: number;
  avatarStage: UserProfileAvatarStage;
  monthlyIncome: number;
  savingsGoal: number;
  joinedAt: string;
}

export interface DashboardSummary {
  balance: number;
  futureScore: number;
  streak: number;
  xp: number;
  level: number;
  monthlySpend: number;
  monthlySavings: number;
  savingsRate: number;
  topSpendCategory: string;
  pendingQuestsCount: number;
  completedQuestsCount: number;
  scoreChange: number;
  avatarStage: string;
}

export interface SpendingChartPoint {
  date: string;
  amount: number;
  category: string;
}

export interface SpendingChart {
  points: SpendingChartPoint[];
  totalThisWeek: number;
  totalLastWeek: number;
  weekOverWeekChange: number;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  debit: "debit",
  credit: "credit",
} as const;

export interface Transaction {
  id: string;
  description: string;
  amount: number;
  category: string;
  type: TransactionType;
  date: string;
  impactScore: number;
}

export type QuestCategory = (typeof QuestCategory)[keyof typeof QuestCategory];

export const QuestCategory = {
  saving: "saving",
  spending: "spending",
  streak: "streak",
  challenge: "challenge",
} as const;

export type QuestDifficulty =
  (typeof QuestDifficulty)[keyof typeof QuestDifficulty];

export const QuestDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface Quest {
  id: string;
  title: string;
  description: string;
  xpReward: number;
  progress: number;
  target: number;
  completed: boolean;
  category: QuestCategory;
  difficulty: QuestDifficulty;
  expiresAt: string;
}

export interface QuestCompletionResult {
  success: boolean;
  xpEarned: number;
  newLevel: number;
  newXp: number;
  leveledUp: boolean;
  message: string;
}

export interface SimulationRequest {
  description: string;
  amount: number;
  category: string;
}

export interface SimulationScenario {
  futureBalance: number;
  futureScore: number;
  scoreImpact: number;
  savingsIn3Months: number;
  savingsIn1Year: number;
  message: string;
  avatarStage: string;
  milestones: string[];
}

export type SimulationResultRiskLevel =
  (typeof SimulationResultRiskLevel)[keyof typeof SimulationResultRiskLevel];

export const SimulationResultRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface SimulationResult {
  disciplined: SimulationScenario;
  chaotic: SimulationScenario;
  riskLevel: SimulationResultRiskLevel;
  isBossBattle: boolean;
}

export type FutureMessageFrom =
  (typeof FutureMessageFrom)[keyof typeof FutureMessageFrom];

export const FutureMessageFrom = {
  disciplined: "disciplined",
  chaotic: "chaotic",
  neutral: "neutral",
} as const;

export type FutureMessageTone =
  (typeof FutureMessageTone)[keyof typeof FutureMessageTone];

export const FutureMessageTone = {
  encouragement: "encouragement",
  warning: "warning",
  celebration: "celebration",
  urgent: "urgent",
} as const;

export interface FutureMessage {
  id: string;
  from: FutureMessageFrom;
  subject: string;
  body: string;
  tone: FutureMessageTone;
  date: string;
  read: boolean;
}

export type BossBattleTier =
  (typeof BossBattleTier)[keyof typeof BossBattleTier];

export const BossBattleTier = {
  miniboss: "miniboss",
  boss: "boss",
  megaboss: "megaboss",
} as const;

export interface BossBattle {
  id: string;
  name: string;
  description: string;
  amount: number;
  category: string;
  scoreRisk: number;
  savingsImpact: number;
  tier: BossBattleTier;
  active: boolean;
}

export interface BossFightResult {
  victory: boolean;
  message: string;
  scoreChange: number;
  xpChange: number;
  newScore: number;
}

export type ListTransactionsParams = {
  limit?: number;
};