# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CashCraft API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  balance: number;
  futureScore: number;
  streak: number;
  xp: number;
  level: number;
  monthlySpend: number;
  monthlySavings: number;
  savingsRate: number;
  topSpendCategory: string;
  pendingQuestsCount: number;
  completedQuestsCount: number;
  scoreChange: number;
  avatarStage: string;
}