# lib/api-zod/src/generated/types/futureMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CashCraft API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FutureMessageFrom } from "./futureMessageFrom";
import type { FutureMessageTone } from "./futureMessageTone";

export interface FutureMessage {
  id: string;
  from: FutureMessageFrom;
  subject: string;
  body: string;
  tone: FutureMessageTone;
  date: Date;
  read: boolean;
}