# lib/api-zod/src/generated/types/simulationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CashCraft API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SimulationResultRiskLevel } from "./simulationResultRiskLevel";
import type { SimulationScenario } from "./simulationScenario";

export interface SimulationResult {
  disciplined: SimulationScenario;
  chaotic: SimulationScenario;
  riskLevel: SimulationResultRiskLevel;
  isBossBattle: boolean;
}