# lib/api-zod/src/generated/types/simulationScenario.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CashCraft API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SimulationScenario {
  futureBalance: number;
  futureScore: number;
  scoreImpact: number;
  savingsIn3Months: number;
  savingsIn1Year: number;
  message: string;
  avatarStage: string;
  milestones: string[];
}